pyproject.toml

[tool.poetry]
name = "DrfPydanticBridge"
version = "0.1.0"
description = "A package for integrating Pydantic with Django and DRF."
authors = ["Your Name <your_email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.0"
djangorestframework = "^3.10"
pydantic = "^1.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
